feat: add methods that count messages in the specified thread & extend int to long in ThreadTopicDTO

+++ src/main/java/dev/yawkar/eternity/web/dto/response/ThreadTopicDTO.java
@@ -12,5 +12,5 @@ public class ThreadTopicDTO {
     private long id;
     private String title;
     private String description;
-    private int messagesNumber;
+    private long messagesNumber;
 }
